backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-cors==1.7.0
websockets==12.0
asyncpg==0.29.0
alembic==1.13.1
jupyter==1.0.0
nbformat==5.9.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
